move mod-specific FirstRun logic to config/module
